requirements.txt
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
autopep8==2.3.2
billiard==4.2.1
binaryornot==0.4.4
boto3==1.35.76
botocore==1.35.99
celery==5.5.3
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickhouse-driver==0.2.9
colorama==0.4.6
cookiecutter==2.6.0
coverage==7.6.9
decorator==5.2.1
dj-database-url==2.3.0
Django==5.1.5
django-cleanup==9.0.0
django-configurations==2.5.1
django-cors-headers==4.7.0
django-filter==24.3
django-js-asset==3.1.2
django-model-utils==5.0.0
django-mptt==0.17.0
django-silk==5.3.2
django-storages==1.14.4
django-unique-upload==0.2.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.3.1
executing==2.2.0
factory_boy==3.3.1
Faker==37.0.0
flake8==7.1.1
ghp-import==2.1.0
gprof2dot==2024.6.6
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mock==5.1.0
newrelic==10.3.1
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-strtobool==1.0.3
pytz==2025.2
PyYAML==6.0.2
pyyaml_env_tag==0.1
redis==6.2.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
text-unidecode==1.3
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
whitenoise==6.9.0
